+++ frontsite/src/app/components/formLogin.tsx
@@ -28,7 +28,7 @@ const FormLogin: React.FC<boxProps> = () => {
                     <Field id="password" type="password" name="password" placeholder="password:" className="w-full h-10 bg-lime-400 rounded-lg p-3"></Field>
                     <div className="w-full h-12 flex justify-between">
                         <button className= "h-full w-1/3 bg-lime-400 rounded-xl" type="submit">Entrar</button>
-                        <a href="./registerPage" className= "h-full w-1/3 bg-lime-400 rounded-xl text-center content-center" type="button">
+                        <a href="./register" className= "h-full w-1/3 bg-lime-400 rounded-xl text-center content-center" type="button">
                             registrar
                         </a>
                     </div>
