+++ frontsite/src/app/components/teacherIcon.tsx
@@ -1,15 +1,15 @@
 import React from "react";
 
 interface teachIcon {
-    id: number
+    teacherID: number
     nome: string,
     foto: string
     estilo: string
 }
 
-const TeacherIcon: React.FC<teachIcon> = ({nome, foto, estilo, id}) => {
+const TeacherIcon: React.FC<teachIcon> = ({nome, foto, estilo, teacherID}) => {
     return(
-        <a href={"/teacherPage/"+id} className={estilo}>
+        <a href={"./teacher/" + teacherID} className={estilo}>
             <div className="h-4/5 w-full overflow-hidden rounded-lg">
                 <img className="w-full rounded-lg" src={foto} alt={foto} />
             </div>
