+++ frontsite/src/app/noLog/login/page.tsx
@@ -1,6 +1,5 @@
 import React, { ChangeEventHandler, MouseEventHandler } from 'react';
-import FormLogin from '../components/formLogin';
-
+import FormLogin from '@/app/components/formLogin';
 
 const loginPage = () => {
